requirements.txt
google-generativeai
langchain-huggingface
langchain-community
chromadb
sentence-transformers
python-dotenv
torch